+++ test/test.txt
@@ -1,2 +1,3 @@
+proc_simple_stats.jl
 problems/proc_gaussian_least_squares.jl
 optimization_routines/proc_barzilai_borwein_gd.jl
